+++ src/main/java/de/maxhenkel/replayvoicechat/net/LocationalSoundPacket.java
@@ -14,10 +14,12 @@ public class LocationalSoundPacket extends AbstractSoundPacket<LocationalSoundPa
     public static ResourceLocation ID = new ResourceLocation(ReplayVoicechat.MOD_ID, "locational_sound");
 
     private Position location;
+    private float distance;
 
-    public LocationalSoundPacket(UUID id, short[] rawAudio, Position location) {
+    public LocationalSoundPacket(UUID id, short[] rawAudio, Position location, float distance) {
         super(id, rawAudio);
         this.location = location;
+        this.distance = distance;
     }
 
     public LocationalSoundPacket() {
@@ -26,6 +28,10 @@ public class LocationalSoundPacket extends AbstractSoundPacket<LocationalSoundPa
 
     public Position getLocation() {
         return location;
+    }
+
+    public float getDistance() {
+        return distance;
     }
 
     @Override
@@ -37,6 +43,11 @@ public class LocationalSoundPacket extends AbstractSoundPacket<LocationalSoundPa
     public LocationalSoundPacket fromBytes(FriendlyByteBuf buf) throws VersionCompatibilityException {
         super.fromBytes(buf);
         location = ReplayVoicechatPlugin.CLIENT_API.createPosition(buf.readDouble(), buf.readDouble(), buf.readDouble());
+        if (version >= 1) {
+            distance = buf.readFloat();
+        } else {
+            distance = (float) ReplayVoicechatPlugin.CLIENT_API.getVoiceChatDistance();
+        }
         return this;
     }
 
@@ -46,6 +57,7 @@ public class LocationalSoundPacket extends AbstractSoundPacket<LocationalSoundPa
         buf.writeDouble(location.getX());
         buf.writeDouble(location.getY());
         buf.writeDouble(location.getZ());
+        buf.writeFloat(distance);
     }
 
     @Override