+++ src/main/java/de/maxhenkel/replayvoicechat/recording/VoicechatRecorder.java
@@ -5,11 +5,9 @@ import de.maxhenkel.replayvoicechat.net.EntitySoundPacket;
 import de.maxhenkel.replayvoicechat.net.LocationalSoundPacket;
 import de.maxhenkel.replayvoicechat.net.Packet;
 import de.maxhenkel.replayvoicechat.net.StaticSoundPacket;
-import de.maxhenkel.voicechat.VoicechatClient;
 import de.maxhenkel.voicechat.api.events.ClientReceiveSoundEvent;
 import de.maxhenkel.voicechat.api.events.ClientSoundEvent;
 import de.maxhenkel.voicechat.voice.client.ClientManager;
-import de.maxhenkel.voicechat.voice.client.ClientPlayerStateManager;
 import io.netty.buffer.Unpooled;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.FriendlyByteBuf;
@@ -22,11 +20,11 @@ public class VoicechatRecorder {
     private static final Minecraft MC = Minecraft.getInstance();
 
     public static void onEntitySound(ClientReceiveSoundEvent.EntitySound event) {
-        send(new EntitySoundPacket(event.getId(), event.getRawAudio(), event.isWhispering()));
+        send(new EntitySoundPacket(event.getId(), event.getRawAudio(), event.isWhispering(), event.getDistance()));
     }
 
     public static void onLocationalSound(ClientReceiveSoundEvent.LocationalSound event) {
-        send(new LocationalSoundPacket(event.getId(), event.getRawAudio(), event.getPosition()));
+        send(new LocationalSoundPacket(event.getId(), event.getRawAudio(), event.getPosition(), event.getDistance()));
     }
 
     public static void onStaticSound(ClientReceiveSoundEvent.StaticSound event) {
@@ -43,7 +41,7 @@ public class VoicechatRecorder {
         if (ReplayVoicechatPlugin.CLIENT_API.getGroup() != null) {
             send(new StaticSoundPacket(id, rawAudio));
         } else {
-            send(new EntitySoundPacket(id, rawAudio, event.isWhispering()));
+            send(new EntitySoundPacket(id, rawAudio, event.isWhispering(), (float) ReplayVoicechatPlugin.CLIENT_API.getVoiceChatDistance()));
         }
     }